experiments/v2_sam3_qwen_vl/requirements.txt
torch>=2.4.0
torchvision>=0.19.0
vllm>=0.6.0
transformers
accelerate
opencv-python
numpy
Pillow
matplotlib
pycocotools
decord
